refactor(user-svc,order-svc): moved addess and it's related tables to user svc
