+++ server/svc/kreditcart-user-svc/src/main/java/com/kreditcart/userservice/Models/City.java
@@ -1,4 +1,4 @@
-package com.kredicart.order.Models;
+package com.kreditcart.userservice.Models;
 
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -7,22 +7,22 @@ import lombok.Setter;
 @Entity
 @Getter
 @Setter
-@Table(name = "countries")
-public class Country extends BaseModel {
+@Table(name = "cities", uniqueConstraints = {
+        @UniqueConstraint(columnNames = {"name", "state_id"})
+})
+public class City extends BaseModel {
     @Column(nullable = false)
     private String name;
-    @Column(nullable = false, unique = true)
-    private String isoCode;
+
+    @ManyToOne
+    @JoinColumn(name = "state_id", nullable = false)
+    private State state;
 
     @PrePersist
     @PreUpdate
     public void capitalizeFields() {
         if (this.name != null) {
             this.name = this.name.toUpperCase();
-        }
-
-        if(this.isoCode != null) {
-            this.isoCode = this.isoCode.toUpperCase();
         }
     }
 }
