+++ server/svc/kreditcart-user-svc/src/main/java/com/kreditcart/userservice/Models/GeoLocation.java
@@ -1,4 +1,4 @@
-package com.kredicart.order.Models;
+package com.kreditcart.userservice.Models;
 
 import jakarta.persistence.Embeddable;
